+++ test/test_graph_filter.py
@@ -60,24 +60,6 @@ class TestGraphFilter:
     - if the edge map is equal to the skeleton
     - if there are no pixels equal to -1 in the edge map
   '''
-  
-  def test_skeletonezer_img_required (self):
-
-    # construct the object
-    skeleton_filter = SkeletonizeImageFilter()
-    
-    # execute the filter without a img
-    with pytest.raises(TypeError):
-      skeleton_filter.Execute()
-
-  def test_skeletonezer_get_without_execute (self):
-
-    # construct the object
-    skeleton_filter = SkeletonizeImageFilter()
-
-    # get the image without the execute
-    with pytest.raises(RuntimeError):
-      skeleton_filter.GetSkeletonImage()
 
   def test_graph_filter_img_required (self):
 
@@ -115,35 +97,6 @@ class TestGraphFilter:
     with pytest.raises(RuntimeError):
       graph_filter.GetEdgeMap()
   
-  def test_ndim_2D_3D (self):
-    
-    # create a 4D image
-    src = np.ones(shape=(10, 10, 10, 10), dtype=np.uint8)
-    src = sitk.GetImageFromArray(src)
-
-    # construct the object
-    skeleton_filter = SkeletonizeImageFilter()
-
-    # execute the filter with a 4D image
-    with pytest.raises(ValueError):
-      skeleton_filter.Execute(src=src)
-
-    # test with 2D input
-    src = np.ones(shape=(10, 10), dtype=np.uint8)
-    src = sitk.GetImageFromArray(src)
-
-    # construct the object
-    skeleton_filter = SkeletonizeImageFilter()
-    skeleton_filter.Execute(src=src)
-
-    # test with 3D input
-    src = np.ones(shape=(10, 10, 10), dtype=np.uint8)
-    src = sitk.GetImageFromArray(src)
-
-    # construct the object
-    skeleton_filter = SkeletonizeImageFilter()
-    skeleton_filter.Execute(src=src)
-
   def test_n_nodes_negative_components (self):
     
     # load the image