+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.0.1] - 2023-09-27
+## [0.0.1] - 2023-09-28
 
 First version of the library.
 This is the starting point of the development of the *pyGraphomics* package.
@@ -33,6 +33,8 @@ Further improvements will occur in the next versions.
 - :computer: [Python] Image resampling to equal spacing for the correct weighing of the network according to the path length
 - :computer: [Python] Test for feature extractor management
 - :computer: [Python] Test for feature extractor configuration file
+- :computer: [Python] Test for image data loader
+- :computer: [Python] Test for image skeletonizer (aka `SkeletonizeImageFilter`)
 - :computer: [Python] Test for `GraphThicknessImageFilter`
 - :computer: [Python] Test for graphomic topological features (aka `GraphomicsTopology`)
 - :computer: [Python] Test for graphomic centrality features (aka `GraphomicsCentrality`)